+++ acoular_cpp/src/trajectory.cpp
@@ -14,15 +14,15 @@ Trajectory::Trajectory() {}
 
 Trajectory::~Trajectory() {}
 
-void Trajectory::set_points(std::map<float, std::vector<float>> points) {
+void Trajectory::set_points(std::map<double, std::vector<double>> points) {
     points_ = points;
-    std::vector<float> t;
+    std::vector<double> t;
     for (auto const& [key, val] : points_) {
         t.push_back(key);
     }
     std::sort(t.begin(), t.end()); // 对时间戳进行排序
     interval_ = std::make_pair(t.front(), t.back()); // 记录时间间隔
-    std::vector<float> xp;
+    std::vector<double> xp;
     for (auto const& [key, val] : points_) {
         xp.insert(xp.end(), val.begin(), val.end());
     }
@@ -36,26 +36,26 @@ void Trajectory::set_points(std::map<float, std::vector<float>> points) {
     }
 }
 
-std::pair<float, float> Trajectory::get_interval() const {
+std::pair<double, double> Trajectory::get_interval() const {
     return interval_; // 返回时间间隔
 }
 
-std::vector<float> Trajectory::location(float t, int der) const {
-    std::vector<float> res(3); // 存储位置或导数
+std::vector<double> Trajectory::location(double t, int der) const {
+    std::vector<double> res(3); // 存储位置或导数
     int n = 1;
     splint_(&tck_[0], &t, &res[0], &n, &der); // 计算位置或导数
     return res;
 }
 
-std::vector<std::vector<float>> Trajectory::traj(float t_start, float t_end, float delta_t, int der) const {
+std::vector<std::vector<double>> Trajectory::traj(double t_start, double t_end, double delta_t, int der) const {
     if (delta_t <= 0) {
         throw std::invalid_argument("delta_t must be positive"); // 如果delta_t小于等于0，则抛出异常
     }
     if (t_start > t_end) {
         throw std::invalid_argument("t_start must be less than or equal to t_end"); // 如果t_start大于t_end，则抛出异常
     }
-    std::vector<std::vector<float>> res; // 存储轨迹
-    for (float t = t_start; t <= t_end; t += delta_t) {
+    std::vector<std::vector<double>> res; // 存储轨迹
+    for (double t = t_start; t <= t_end; t += delta_t) {
         res.push_back(location(t, der)); // 计算位置或导数，并将其添加到轨迹中
     }
     return res;
@@ -63,7 +63,7 @@ std::vector<std::vector<float>> Trajectory::traj(float t_start, float t_end, flo
 
 
 // TODO ： 未完成
-void Trajectory::splprep_(float *t, float *x, int *k, int *s, float *tck, int *info) const
+void Trajectory::splprep_(double *t, double *x, int *k, int *s, double *tck, int *info) const
 {
     int n = (*k) + 1; // 样本点的数量
     int p = 3; // 样条插值的阶数
@@ -78,16 +78,16 @@ void Trajectory::splprep_(float *t, float *x, int *k, int *s, float *tck, int *i
     Eigen::ArrayXf h = (T.tail(n - 1) - T.head(n - 1)).matrix();
 
     // 构造三对角矩阵
-    Eigen::SparseMatrix<float> A(m, m);
-    std::vector<Eigen::Triplet<float>> triplets;
+    Eigen::SparseMatrix<double> A(m, m);
+    std::vector<Eigen::Triplet<double>> triplets;
     triplets.reserve(3 * m - 2);
-    triplets.push_back(Eigen::Triplet<float>(0, 0, 1.0));
+    triplets.push_back(Eigen::Triplet<double>(0, 0, 1.0));
     for (int i = 1; i < n - 1; i++) {
-        triplets.push_back(Eigen::Triplet<float>(i, i - 1, h[i - 1]));
-        triplets.push_back(Eigen::Triplet<float>(i, i, 2.0 * (h[i - 1] + h[i])));
-        triplets.push_back(Eigen::Triplet<float>(i, i + 1, h[i]));
+        triplets.push_back(Eigen::Triplet<double>(i, i - 1, h[i - 1]));
+        triplets.push_back(Eigen::Triplet<double>(i, i, 2.0 * (h[i - 1] + h[i])));
+        triplets.push_back(Eigen::Triplet<double>(i, i + 1, h[i]));
     }
-    triplets.push_back(Eigen::Triplet<float>(n - 1, n - 1, 1.0));
+    triplets.push_back(Eigen::Triplet<double>(n - 1, n - 1, 1.0));
     A.setFromTriplets(triplets.begin(), triplets.end());
 
     // 构造右侧向量
@@ -99,7 +99,7 @@ void Trajectory::splprep_(float *t, float *x, int *k, int *s, float *tck, int *i
     }
 
     // 解线性方程组
-    Eigen::SparseLU<Eigen::SparseMatrix<float>> solver;
+    Eigen::SparseLU<Eigen::SparseMatrix<double>> solver;
     solver.compute(A);
     if (solver.info() != Eigen::Success) {
         *info = 1; // 矩阵分解失败
@@ -111,9 +111,9 @@ void Trajectory::splprep_(float *t, float *x, int *k, int *s, float *tck, int *i
     // 计算TCK系数
     int idx = 0;
     for (int i = 0; i < n - 1; i++) {
-        float t1 = T[i];
-        float t2 = T[i + 1];
-        float h = t2 - t1;
+        double t1 = T[i];
+        double t2 = T[i + 1];
+        double h = t2 - t1;
         Eigen::ArrayXf c1 = C.row(i);
         Eigen::ArrayXf c2 = C.row(i + 1);
         Eigen::ArrayXf d1 = (c2 - c1) / (3.0 * h);
@@ -132,7 +132,7 @@ void Trajectory::splprep_(float *t, float *x, int *k, int *s, float *tck, int *i
     *info = 0;
 }
 
-void Trajectory::splint_(float *tck, float *t, float *res, int *n, int *der) const
+void Trajectory::splint_(double *tck, double *t, double *res, int *n, int *der) const
 {
     int m = (*n) / 3; // 样本点的数量
     int k = 3; // 样条插值的阶数
@@ -147,13 +147,13 @@ void Trajectory::splint_(float *tck, float *t, float *res, int *n, int *der) con
     int d = *der;
     Eigen::ArrayXXf R(m, d == 0 ? 3 : 1);
     for (int i = 0; i < m; i++) {
-        float u = U[i];
+        double u = U[i];
         int j = std::min(int(u * (m - 1)), m - 2);
-        float h = T[j + 1] - T[j];
-        float a = (T[j + 1] - u) / h;
-        float b = (u - T[j]) / h;
-        float c = (1.0 / 6.0) * (a * a * a - a) * h * h;
-        float d = (1.0 / 6.0) * (b * b * b - b) * h * h;
+        double h = T[j + 1] - T[j];
+        double a = (T[j + 1] - u) / h;
+        double b = (u - T[j]) / h;
+        double c = (1.0 / 6.0) * (a * a * a - a) * h * h;
+        double d = (1.0 / 6.0) * (b * b * b - b) * h * h;
         Eigen::ArrayXf C = TCK.segment<10>(10 * j);
         if (d == 0) {
             R.row(i) = C.segment<3>(0) + a * C.segment<3>(3) + c * C.segment<3>(6) + d * C.segment<3>(9);
